refactor(codereview): change the usage of startReview
