+++ src/core/costrict/code-review/index.ts
@@ -1,80 +1,15 @@
 import * as vscode from "vscode"
-import path from "node:path"
 
 import { ClineProvider } from "../../webview/ClineProvider"
 import { getCommand } from "../../../utils/commands"
 import { toRelativePath } from "../../../utils/path"
 import { CostrictCommandId } from "@roo-code/types"
-import { t } from "../../../i18n"
-import { IssueStatus, TaskStatus, ReviewTarget, ReviewTargetType } from "../../../shared/codeReview"
+import { IssueStatus, ReviewTarget, ReviewTargetType } from "../../../shared/codeReview"
 import { getVisibleProviderOrLog } from "../../../activate/registerCommands"
 
 import { CodeReviewService } from "./codeReviewService"
 import { CommentService } from "../../../integrations/comment"
 import type { ReviewComment } from "./reviewComment"
-import { zgsmCodebaseIndexManager } from "../codebase-index"
-const startReview = async (
-	reviewInstance: CodeReviewService,
-	targets: ReviewTarget[],
-	isReviewRepo: boolean = false,
-) => {
-	const visibleProvider = await ClineProvider.getInstance()
-	const codebaseSyncService = zgsmCodebaseIndexManager
-	if (visibleProvider) {
-		const filePaths = targets.map((target) => path.join(visibleProvider.cwd, target.file_path))
-		const { zgsmCodebaseIndexEnabled, apiConfiguration } = await visibleProvider.getState()
-		if (apiConfiguration.apiProvider !== "zgsm") {
-			vscode.window.showInformationMessage(t("common:review.tip.api_provider_not_support"))
-			return
-		}
-		reviewInstance.setProvider(visibleProvider)
-		if (!isReviewRepo) {
-			try {
-				const success = await vscode.window.withProgress(
-					{
-						location: vscode.ProgressLocation.Notification,
-						title: t("common:review.tip.file_check"),
-					},
-					async (progress) => {
-						const workspace = visibleProvider.cwd
-						const { success } = await codebaseSyncService.checkIgnoreFiles({
-							workspacePath: workspace,
-							workspaceName: path.basename(workspace),
-							filePaths,
-						})
-						progress.report({ increment: 100 })
-						return success
-					},
-				)
-				if (!success) {
-					vscode.window.showInformationMessage(t("common:review.tip.codebase_sync_ignore_file"))
-					return
-				}
-			} catch (error) {
-				vscode.window.showInformationMessage(t("common:review.tip.service_unavailable"))
-				return
-			}
-		}
-		const res = await codebaseSyncService.getIndexStatus(visibleProvider.cwd)
-		const { codegraph } = res.data
-		visibleProvider.postMessageToWebview({
-			type: "reviewPagePayload",
-			payload: {
-				targets,
-				isCodebaseReady:
-					zgsmCodebaseIndexEnabled && (codegraph.status === "success" || codegraph.process === 100),
-			},
-		})
-		reviewInstance.resetStatus()
-		visibleProvider.postMessageToWebview({
-			type: "action",
-			action: "codeReviewButtonClicked",
-		})
-		if (zgsmCodebaseIndexEnabled && (codegraph.status === "success" || codegraph.process === 100)) {
-			await reviewInstance.startReviewTask(targets)
-		}
-	}
-}
 export function initCodeReview(
 	context: vscode.ExtensionContext,
 	provider: ClineProvider,
@@ -103,7 +38,8 @@ export function initCodeReview(
 			const fileUri = editor.document.uri
 			const range = editor.selection
 			const cwd = visibleProvider.cwd.toPosix()
-			startReview(reviewInstance, [
+			reviewInstance.setProvider(visibleProvider)
+			reviewInstance.startReview([
 				{
 					type: ReviewTargetType.CODE,
 					file_path: toRelativePath(fileUri.fsPath.toPosix(), cwd),
@@ -126,15 +62,16 @@ export function initCodeReview(
 					}
 				}),
 			)
-			startReview(reviewInstance, targets)
+			reviewInstance.setProvider(visibleProvider)
+			reviewInstance.startReview(targets)
 		},
 		reviewRepo: async () => {
 			const visibleProvider = await ClineProvider.getInstance()
 			if (!visibleProvider) {
 				return
 			}
-			startReview(
-				reviewInstance,
+			reviewInstance.setProvider(visibleProvider)
+			reviewInstance.startReview(
 				[
 					{
 						type: ReviewTargetType.FOLDER,
@@ -173,4 +110,4 @@ export function initCodeReview(
 	}
 }
 
-export { CodeReviewService, startReview, ReviewTargetType }
+export { CodeReviewService, ReviewTargetType }
